sizeof

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ int main(void)
 {
 	printf("Size of a char: %lu byte(S)\n", sizeof(char));
 	printf("Size of an int: %lu byte(S)\n", sizeof(int));
-	printf("Size of a long int: %lu byte(S)\n", sizeof(long int));
+	printf("Size of a long int: %lu byte(S)\n", sizeof(long long int));
 	printf("Size of a long long int: %lu byte(S)\n", sizeof(long long int));
 	printf("Size of a float: %lu byte(s)\n", sizeof(float));
 		return (0);
